siso: fix build error on 32bit arch
https://logs.chromium.org/logs/infra-internal/buildbucket/cr-buildbucket/8770902735481898881/+/u/l_GOOS:linux_GOARCH:386_cipd_-_build_packages/stdout

 # infra/build/siso/execute/localexec
 build/siso/execute/localexec/rusage_unix.go:19:12: cannot use u.Maxrss (variable of type int32) as int64 value in struct literal

always convert to int64

Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4838202
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#59461}
GitOrigin-RevId: c74271d73fc05c1bc4750ad7f64bfeaf38478aa6

+++ execute/localexec/rusage_unix.go
@@ -16,7 +16,8 @@ import (
 func rusage(cmd *exec.Cmd) *epb.Rusage {
 	if u, ok := cmd.ProcessState.SysUsage().(*syscall.Rusage); ok {
 		return &epb.Rusage{
-			MaxRss: u.Maxrss,
+			// 32bit arch may use int32 for Maxrss.
+			MaxRss: int64(u.Maxrss),
 		}
 	}
 	return nil
